+++ Runtime/ConvertFlags.cs
@@ -6,7 +6,11 @@ namespace UnderLogic.Serialization.Toml
     internal enum ConvertFlags
     {
         None,
-        ForceArray = 0x1,
-        ForceList = 0x2,
+        Literal = 0x1,
+        Multiline = 0x2,
+        ForceArray = 0x4,
+        ForceList = 0x8,
+        ForceInline = 0x10,
+        ForceExpand = 0x20
     }
 }