+++ Runtime/TomlCasingAttribute.cs
@@ -3,7 +3,7 @@ using System;
 namespace UnderLogic.Serialization.Toml
 {
     [AttributeUsage(AttributeTargets.Class | AttributeTargets.Struct | AttributeTargets.Field)]
-    public class TomlCasingAttribute : Attribute
+    public sealed class TomlCasingAttribute : Attribute
     {
         public StringCasing Casing { get; }