+++ CHANGELOG.md
@@ -4,6 +4,20 @@ All notable changes to this library will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.1] - 2022-12-30
+
+### Added
+
+- `TomlLiteralAttribute` custom attribute for serializing literal strings
+- `TomlMultilineAttribute` custom attribute for serializing multi-line strings (including literals)
+- `TomlInlineAttribute` custom attribute for serializing inline tables explicitly
+- `TomlExpandAttribute` custom attribute for serializing expanded tables explicitly
+- More unit tests for various serialization cases
+
+### Changed
+
+- `TomlFloat` values can be coerced into integer values (will be truncated)
+
 ## [0.4.0] - 2022-12-30
 
 ### Added